Fix projectile Hit Advantage
Two issues:
1. The hitlag was being calculated with the post-1v1 multiplier damage
2. It doesn't take into account the fact that X frames of projectile hitlag only means X-1 additional advantage frames.

As an example of (2), consider a move that hits on frame 10 and deals 5 frames of hitlag. If it's a direct attack, the victim starts hitstun on the attacker's frame 11. If it's a projectile, the victim is in hitlag on the attacker's frames 10-14 and starts hitstun on frame 15 - this is only 4 frames later.