Fix Sora nair/fair 2/3 landing lag
This was leading to misleading frame advantage on the calculator, as the landing lag for some of Sora's aerials was wrong.

Sora's aerial landing lag:
Nair 1: 9
Nair 2: 10
Nair 3: 11
Fair 1: 12
Fair 2: 13
Fair 3: 14

The params for each character include NairLandingLag/FairLandingLag each as a single number, which isn't sufficient for Sora. I thus added an optional custom_landing_lag param for a move, and set it for Nair 2/3 and Fair 2/3.